Exclude Windows-only code from source build
