+++ src/Features/Core/Portable/SymbolSearch/Windows/IAddReferenceDatabaseWrapper.cs
