+++ src/Features/Core/Portable/SymbolSearch/Windows/IDatabaseFactoryService.cs
