+++ src/Features/Core/Portable/SymbolSearch/Windows/IDelayService.cs
