+++ src/Features/Core/Portable/SymbolSearch/Windows/IIOService.cs
