+++ src/Features/Core/Portable/SymbolSearch/Windows/IPatchService.cs
