+++ src/Features/Core/Portable/SymbolSearch/Windows/IRemoteControlService.cs
