+++ src/Features/Core/Portable/SymbolSearch/Windows/SymbolSearchUpdateEngine.PatchService.cs
@@ -11,7 +11,7 @@ namespace Microsoft.CodeAnalysis.SymbolSearch
         private class PatchService : IPatchService
         {
             public byte[] ApplyPatch(byte[] databaseBytes, byte[] patchBytes)
-                => Patching.Delta.ApplyPatch(databaseBytes, patchBytes);
+                => NativePatching.ApplyPatch(databaseBytes, patchBytes);
         }
     }
 }
