+++ src/Features/Core/Portable/SymbolSearch/Windows/SymbolSearchUpdateEngine.RemoteControlService.cs
