+++ src/Features/Core/Portable/SymbolSearch/Windows/SymbolSearchUpdateEngine.cs
